Add files via upload
In the above example, the Cat class extends the Animal class and overrides its makeSound() method. The super.makeSound() call in the makeSound() method of the Cat class is used to call the makeSound() method of the superclass (Animal), allowing the Cat class to inherit and extend the behavior of the Animal class.